fix alloc

+++ src/alloc/error.rs
@@ -7,7 +7,7 @@ pub enum Error {
     Necronomicon(#[from] necronomicon::Error),
 
     #[error("out of memory {used}/{total}")]
-    OutOfMemory { total: u64, used: usize },
+    OutOfMemory { total: u64, used: u64 },
 
     #[error("bad node {0}")]
     BadNode(usize),
